requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
mcp==1.13.1
    # via -r requirements.in
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
pydantic==2.11.7
    # via
    #   fastapi
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.1.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.47.3
    # via
    #   fastapi
    #   mcp
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.35.0
    # via
    #   -r requirements.in
    #   mcp
